Minor updates to classes due to double vs int data presentation bug found.

+++ src/main/java/repository/MondasRepository.java
@@ -92,7 +92,7 @@ public class MondasRepository {
             if (!resultSet.next()) break;
             Mondas aplicacao = new Mondas(
                     resultSet.getInt("OPERACAOID"),
-                    resultSet.getInt("quantidade"),
+                    resultSet.getDouble("quantidade"),
                     resultSet.getString("UNIDADEMEDIDA")
             );
             mondasList.add(aplicacao);
